server/requirements.txt
fastapi
uvicorn
requests
Pillow
pydantic